+++ monarch/lib/src/monarch_app/ready_signal.dart
@@ -7,7 +7,10 @@ class ReadySignal {
   final _changeStreamController = StreamController<bool>.broadcast();
   Stream<bool> get changeStream => _changeStreamController.stream;
 
-  void starting() => _isReady = false;
+  void loading() {
+    _isReady = false;
+    _changeStreamController.add(_isReady);
+  }
 
   void ready() {
     _isReady = true;