+++ monarch/lib/src/monarch_app/start_monarch.dart
@@ -4,11 +4,10 @@ import 'package:stack_trace/stack_trace.dart';
 import 'package:monarch_utils/log.dart';
 import 'package:monarch_utils/log_config.dart';
 
+import 'monarch_data_instance.dart';
 import 'ready_signal.dart';
 import 'device_definitions.dart';
 import 'story_scale_definitions.dart';
-import 'localizations_delegate_loader.dart';
-import 'active_locale.dart';
 import 'active_theme.dart';
 import 'channel_methods_sender.dart';
 import 'channel_methods_receiver.dart';
@@ -16,55 +15,42 @@ import 'standard_themes.dart';
 import 'stories_errors.dart';
 import 'story_app.dart';
 import 'monarch_data.dart';
-import 'user_message.dart';
 import 'vm_service_client.dart';
 import 'monarch_binding.dart';
 
 final _logger = Logger('Start');
 
-void startMonarch(
-    String packageName,
-    List<MetaLocalization> userMetaLocalizations,
-    List<MetaTheme> userMetaThemes,
-    Map<String, MetaStories> metaStoriesMap) {
+void startMonarch(MonarchData Function() getMonarchData) {
   Chain.capture(() {
-    _startMonarch(
-        packageName, userMetaLocalizations, userMetaThemes, metaStoriesMap);
+    _startMonarch(getMonarchData);
   }, onError: handleUncaughtError);
 }
 
-void _startMonarch(
-    String packageName,
-    List<MetaLocalization> userMetaLocalizations,
-    List<MetaTheme> userMetaThemes,
-    Map<String, MetaStories> metaStoriesMap) async {
+void _startMonarch(MonarchData Function() getMonarchData) async {
   final monarchBinding = MonarchBinding.ensureInitialized() as MonarchBinding;
 
   _setUpLog();
+  readySignal.loading();
+  loadMonarchDataInstance(getMonarchData);
+  handleFlutterFrameworkErrors();
 
-  readySignal.starting();
-  _logger.finest('Starting Monarch flutter app');
+  monarchBinding.reassembleCallback = () async {
+    loadMonarchDataInstance(getMonarchData);
+    await channelMethodsSender.sendMonarchData(monarchDataInstance);
+  };
 
-  userMetaLocalizations =
-      _validateAndFilterMetaLocalizations(userMetaLocalizations);
-  userMetaThemes = _validateAndFilterMetaThemes(userMetaThemes);
-
-  final monarchData = MonarchData(
-      packageName, userMetaLocalizations, userMetaThemes, metaStoriesMap);
-
-  handleFlutterFrameworkErrors(monarchData);
-  activeTheme.setMetaThemes([...userMetaThemes, ...standardMetaThemes]);
-  activeLocale =
-      ActiveLocale(LocalizationsDelegateLoader(monarchData.metaLocalizations));
-
-  monarchBinding.attachRootWidget(MonarchStoryApp(
-    monarchData: monarchData,
-  ));
+  monarchBinding.attachRootWidget(MonarchStoryApp());
   monarchBinding.scheduleFrame();
 
   receiveChannelMethodCalls();
   await _connectToVmService();
-  _sendInitialChannelMethodCalls(monarchData);
+  await _sendDefinitions();
+  await channelMethodsSender.sendMonarchData(monarchDataInstance);
+  await channelMethodsSender.sendReadySignal();
+}
+
+void _setUpLog() {
+  writeLogEntryStream(print, printTimestamp: false, printLoggerName: true);
 }
 
 Future<void> _connectToVmService() async {
@@ -79,52 +65,10 @@ Future<void> _connectToVmService() async {
   }
 }
 
-void _setUpLog() {
-  writeLogEntryStream(print, printTimestamp: false, printLoggerName: true);
-}
-
-List<MetaLocalization> _validateAndFilterMetaLocalizations(
-    List<MetaLocalization> metaLocalizationList) {
-  final _list = <MetaLocalization>[];
-  for (var item in metaLocalizationList) {
-    if (item.delegate == null) {
-      printUserMessage(
-          'Info: ${item.delegateClassName} doesn\'t extend LocalizationsDelegate<T>. '
-          'It will be ignored.');
-    } else if (item.locales.isEmpty) {
-      printUserMessage(
-          'Info: @MonarchLocalizations annotation on ${item.delegateClassName} '
-          'doesn\'t declare any locales. It will be ignored.');
-    } else {
-      _logger.fine(
-          'Valid localization found on class ${item.delegateClassName} with '
-          'annotated locales: ${item.locales.map((e) => e.languageCode).toList()}');
-      _list.add(item);
-    }
-  }
-  return _list;
-}
-
-List<MetaTheme> _validateAndFilterMetaThemes(List<MetaTheme> metaThemeList) {
-  final _list = <MetaTheme>[];
-  for (var item in metaThemeList) {
-    if (item.theme == null) {
-      printUserMessage(
-          'Info: Theme "${item.name}" is not of type ThemeData. It will be ignored.');
-    } else {
-      _logger.fine('Valid theme found: ${item.name}');
-      _list.add(item);
-    }
-  }
-  return _list;
-}
-
-void _sendInitialChannelMethodCalls(MonarchData monarchData) async {
+Future<void> _sendDefinitions() async {
   await channelMethodsSender.sendPing();
   await channelMethodsSender.sendDefaultTheme(activeTheme.defaultMetaTheme.id);
   await channelMethodsSender.sendDeviceDefinitions(DeviceDefinitions());
   await channelMethodsSender.sendStoryScaleDefinitions(StoryScaleDefinitions());
   await channelMethodsSender.sendStandardThemes(StandardThemes());
-  await channelMethodsSender.sendMonarchData(monarchData);
-  await channelMethodsSender.sendReadySignal();
 }