+++ monarch/lib/src/monarch_app/story_app.dart
@@ -1,20 +1,18 @@
 import 'dart:async';
 
-import 'package:flutter/widgets.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
 import 'active_locale.dart';
 import 'active_story_scale.dart';
-import 'monarch_data.dart';
 import 'monarch_binding.dart';
+import 'monarch_data_instance.dart';
 import 'ready_signal.dart';
 import 'story_view.dart';
 
 class MonarchStoryApp extends StatefulWidget {
-  final MonarchData monarchData;
-
-  MonarchStoryApp({required this.monarchData});
+  MonarchStoryApp();
 
   @override
   State<StatefulWidget> createState() {
@@ -25,7 +23,7 @@ class MonarchStoryApp extends StatefulWidget {
 class _MonarchStoryAppState extends State<MonarchStoryApp> {
   late bool _isReady;
   late double _storyScale;
-  late LocaleLoadingStatus _localeLoadingStatus;
+  late Locale? _locale;
   final _streamSubscriptions = <StreamSubscription>[];
 
   _MonarchStoryAppState();
@@ -36,15 +34,14 @@ class _MonarchStoryAppState extends State<MonarchStoryApp> {
 
     _isReady = readySignal.isReady;
     _storyScale = activeStoryScale.value;
-    _localeLoadingStatus = activeLocale.loadingStatus;
+    _locale = activeLocale.value;
 
     _streamSubscriptions.addAll([
       readySignal.changeStream
           .listen((isReady) => setState(() => _isReady = isReady)),
       activeStoryScale.stream
           .listen((value) => setState(() => _storyScale = value)),
-      activeLocale.loadingStatusStream
-          .listen((status) => setState(() => _localeLoadingStatus = status))
+      activeLocale.stream.listen((value) => setState(() => _locale = value))
     ]);
   }
 
@@ -60,118 +57,24 @@ class _MonarchStoryAppState extends State<MonarchStoryApp> {
   Widget build(BuildContext context) {
     monarchBindingInstance.lockEventsWhileRendering();
     if (_isReady) {
-      if (widget.monarchData.metaLocalizations.isEmpty) {
-        return MonarchScaleMaterialApp(
-            scale: _storyScale,
-            home: MonarchStoryView(
-              monarchData: widget.monarchData,
-              localeKey: '__NA__',
-            ));
-      } else {
-        return MonarchLocalizedStoryApp(
-            monarchData: widget.monarchData,
-            scale: _storyScale,
-            loadingStatus: _localeLoadingStatus);
-      }
+      return MonarchMaterialApp(
+          scale: _storyScale, locale: _locale, home: MonarchStoryView());
     } else {
-      return MonarchScaleMaterialApp(
+      return MonarchMaterialApp(
           scale: _storyScale,
+          locale: null,
           home: MonarchSimpleMessageView(message: 'Loading...'));
     }
   }
 }
 
-class MonarchScaleMaterialApp extends StatelessWidget {
+class MonarchMaterialApp extends StatelessWidget {
   final double scale;
-  final Widget? home;
-
-  MonarchScaleMaterialApp({Key? key, required this.scale, this.home})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Transform.scale(
-        scale: scale,
-        alignment: Alignment.topLeft,
-        child: MaterialApp(debugShowCheckedModeBanner: false, home: home));
-  }
-}
-
-class MonarchLocalizedStoryApp extends StatelessWidget {
-  final MonarchData monarchData;
-  final double scale;
-  final LocaleLoadingStatus loadingStatus;
-
-  MonarchLocalizedStoryApp(
-      {Key? key,
-      required this.monarchData,
-      required this.scale,
-      required this.loadingStatus})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    switch (loadingStatus) {
-      case LocaleLoadingStatus.initial:
-      case LocaleLoadingStatus.inProgress:
-        return MonarchScaleMaterialApp(
-            scale: scale,
-            home: MonarchSimpleMessageView(message: 'Loading locale...'));
-
-      case LocaleLoadingStatus.done:
-        return _buildOnLocaleLoaded();
-
-      case LocaleLoadingStatus.error:
-        return MonarchScaleMaterialApp(
-            scale: scale,
-            home: MonarchSimpleMessageView(
-                message: 'Unexpected error. Please see console for details.'));
-
-      default:
-        throw 'Unexpected status, got $loadingStatus';
-    }
-  }
-
-  Widget _buildOnLocaleLoaded() {
-    activeLocale.assertIsLoaded();
-    if (activeLocale.canLoad!) {
-      return MonarchLocalizedScaleMaterialApp(
-          scale: scale,
-          localizationsDelegates: [
-            ...monarchData.metaLocalizations.map((x) => x.delegate!),
-            ...GlobalMaterialLocalizations.delegates,
-          ],
-          supportedLocales: monarchData.allLocales,
-          locale: activeLocale.locale,
-          home: MonarchStoryView(
-              monarchData: monarchData,
-              localeKey: activeLocale.locale!.toLanguageTag()));
-    } else {
-      return MonarchScaleMaterialApp(
-          scale: scale,
-          home: MonarchSimpleMessageView(
-              message:
-                  'Error loading locale ${activeLocale.locale!.toLanguageTag()}. '
-                  'Please see console for details.'));
-    }
-  }
-}
-
-class MonarchLocalizedScaleMaterialApp extends StatelessWidget {
-  final double scale;
-  final Widget? home;
-
-  final Iterable<LocalizationsDelegate<dynamic>>? localizationsDelegates;
-  final Iterable<Locale> supportedLocales;
   final Locale? locale;
+  final Widget home;
 
-  MonarchLocalizedScaleMaterialApp(
-      {Key? key,
-      required this.scale,
-      this.home,
-      this.localizationsDelegates,
-      required this.supportedLocales,
-      this.locale})
+  MonarchMaterialApp(
+      {Key? key, required this.scale, required this.locale, required this.home})
       : super(key: key);
 
   @override
@@ -181,9 +84,28 @@ class MonarchLocalizedScaleMaterialApp extends StatelessWidget {
         alignment: Alignment.topLeft,
         child: MaterialApp(
             debugShowCheckedModeBanner: false,
-            home: home,
-            localizationsDelegates: localizationsDelegates,
-            supportedLocales: supportedLocales,
-            locale: locale));
+            scrollBehavior: MonarchScrollBehavior(),
+            localizationsDelegates:
+                locale == null || monarchDataInstance.metaLocalizations.isEmpty
+                    ? null
+                    : [
+                        ...monarchDataInstance.metaLocalizations
+                            .map((x) => x.delegate!),
+                        ...GlobalMaterialLocalizations.delegates,
+                      ],
+            supportedLocales:
+                locale == null || monarchDataInstance.allLocales.isEmpty
+                    ? const <Locale>[Locale('en', 'US')]
+                    : monarchDataInstance.allLocales,
+            locale: locale,
+            home: home));
   }
+}
+
+class MonarchScrollBehavior extends MaterialScrollBehavior {
+  @override
+  Set<PointerDeviceKind> get dragDevices => {
+        PointerDeviceKind.touch,
+        PointerDeviceKind.mouse,
+      };
 }