+++ monarch/CHANGELOG.md
@@ -1,4 +1,20 @@
-## 2.0.0
+## 2.1.0 - 2022-01-11
+- MonarchScrollBehavior to simulate drag gestures with the mouse
+- Reassemble (i.e. hot reload) using MonarchBinding.
+- Reset story channel method. 
+- Generate MonarchData code using getters instead of final static variables.
+- Support getters or const for code elements annotated with 
+  @MonarchLocalizations and @MonarchTheme.
+- Print warnings during code generation and monarch data validation.
+- Remove localizations validation and loader status, any localizations 
+  errors are handled as expected by error handling code.
+- Simplify MaterialApp composition to avoid changing widget tree as 
+  localizations are added or removed.
+- Simplify StoryApp.
+- Simplify MonarchStoryView: use UniqueKey instead of ValueKey. Framework should 
+  always render a fresh story.
+
+## 2.0.0 - 2021-10-19
 - Handle common Flutter error "RenderBox was not laid out" by showing a placeholder
   error view (MonarchStoryErrorView), which prevents secondary errors from 
   polluting the experience.
@@ -16,7 +32,7 @@
 - Simplify Monarch widget tree
 - Use lints instead of pedantic
 
-## 1.1.0
+## 1.1.0 - 2021-10-19
 - All changes from version 2.0.0 but compatible with older Flutter SDK versions 
   (prior to Flutter 2.4.0-4.0.pre)